requirements.txt
pyqt5
pyqtgraph
pyserial
